+++ test/e2e/fixtures/import-api.js
@@ -32,34 +32,32 @@ const postInit = {
 
 // api get
 await res(`${base}/api/get/test`, [200]).then((json) => {
-    if (!(json?.ok && json?.data === "hello"))
-        throw new Error(`json.ok: ${json.ok},json.data: ${json.data},json.code: ${json.code}`);
+    if (json !== "hello") throw new Error(JSON.stringify(json));
 });
 
 await res(`${base}/api/get/not-found`, [404]).then((json) => {
-    if (!(!json?.ok && json?.code === "API_NOT_FOUND" && typeof json?.message === "string"))
-        throw new Error(`json.ok: ${json.ok},json.data: ${json.data},json.code: ${json.code}`);
+    if (!(json?.code === "API_NOT_FOUND" && typeof json?.message === "string"))
+        throw new Error(JSON.stringify(json));
 });
 
 await res(`${base}/api/get/error`, [500]).then((json) => {
-    if (!(!json?.ok && json?.code === "API_INTERNAL_ERROR" && typeof json?.message === "string"))
-        throw new Error(`json.ok: ${json.ok},json.data: ${json.data},json.code: ${json.code}`);
+    if (!(json?.code === "API_INTERNAL_ERROR" && typeof json?.message === "string"))
+        throw new Error(JSON.stringify(json));
 });
 
 // api post
 await res(`${base}/api/post/test`, [200], postInit).then((json) => {
-    if (!(json.ok && json.data === "hello"))
-        throw new Error(`json.ok: ${json.ok},json.data: ${json.data},json.code: ${json.code}`);
+    if (json !== "hello") throw new Error(JSON.stringify(json));
 });
 
 await res(`${base}/api/post/not-found`, [404], postInit).then((json) => {
-    if (!(!json?.ok && json?.code === "API_NOT_FOUND" && typeof json?.message === "string"))
-        throw new Error(`json.ok: ${json.ok},json.data: ${json.data},json.code: ${json.code}`);
+    if (!(json?.code === "API_NOT_FOUND" && typeof json?.message === "string"))
+        throw new Error(JSON.stringify(json));
 });
 
 await res(`${base}/api/post/error`, [500], postInit).then((json) => {
-    if (!(!json?.ok && json?.code === "API_INTERNAL_ERROR" && typeof json?.message === "string"))
-        throw new Error(`json.ok: ${json.ok},json.data: ${json.data},json.code: ${json.code}`);
+    if (!(json?.code === "API_INTERNAL_ERROR" && typeof json?.message === "string"))
+        throw new Error(JSON.stringify(json));
 });
 
 await server.stop();
@@ -67,6 +65,8 @@ await server.stop();
 async function res(url, status, init) {
     const response = await fetch(url, init);
     if (!status.includes(response.status))
+        throw new Error(`response.ok:${response.ok},response.status:${response.status}`);
+    if (response.ok !== status.includes(200))
         throw new Error(`response.ok:${response.ok},response.status:${response.status}`);
     const json = await response.json();
     return json;