+++ test/integration/fixtures/api/api-overwrite.ts
@@ -16,17 +16,16 @@ server.onApi("GET:/get/a", () => {
 await server.start();
 const port = server.getPort();
 
-const res = async (url: string, init?: RequestInit | undefined): Promise<object> => {
+const res = async (url: string, init?: RequestInit | undefined): Promise<unknown> => {
     const response = await fetch(url, init);
-    if (!(response.ok && [404, 200].includes(response.status))) throw new Error();
+    if (!(response.ok && response.status === 200)) throw new Error();
 
     const json = await response.json();
     return json;
 };
 
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-await res(`http://localhost:${port}/api/get/a`).then((json: any) => {
-    if (!(json?.ok && json?.data === "the hello")) throw new Error(JSON.stringify(json));
+await res(`http://localhost:${port}/api/get/a`).then((json) => {
+    if (json !== "the hello") throw new Error(JSON.stringify(json));
 });
 
 await server.stop();